DSSSM-280 #time 4m Resolving integration issues

+++ bash.d/pull_from_dev.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-mv bootstrap_dss_islandora_dev.info bootstrap_lafayette_lib_dss.info
+if [[ -f bootstrap_dss_islandora_dev.info  ]] ; then
+    mv bootstrap_dss_islandora_dev.info bootstrap_lafayette_lib_dss.info
+fi
 
 sed -i 's/bootstrap_dss_islandora_dev/bootstrap_lafayette_lib_dss/' template.php
 sed -i 's/bootstrap_dss_islandora_dev/bootstrap_lafayette_lib_dss/' includes/*inc
